fix CodeQL issue
